Исправил .gitignore. Убрал лишние интерфейсы, теперь можно создавать команду сразу с конечным состоянием и не париться, как оно до него дойдет. Тесты теперь создаются во временной папке с помощью TemporaryFolder.

+++ src/main/java/org/example/command/DeleteCommand.java
@@ -1,52 +1,50 @@
 package org.example.command;
 
 import org.example.internal.ConstantManager;
+import org.example.internal.FileManager;
 import org.example.state.State;
-import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
+import org.example.state.StateSwitcher;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 
-import org.example.internal.FileManager;
-import org.telegram.telegrambots.meta.api.objects.Message;
-
 import java.io.IOException;
+
+import static org.example.state.State.ON_COMMAND_FROM_USER;
 
 /**
  * Команда /delete.
  */
-public class DeleteCommand extends AbstractCommand implements OneStateCommand
+public class DeleteCommand extends AbstractCommand
 {
     private final FileManager fileManager;
 
-    public DeleteCommand()
+    public DeleteCommand(FileManager fileManager)
     {
-        fileManager = new FileManager();
+        super(new StateSwitcher(ON_COMMAND_FROM_USER));
+        this.fileManager = fileManager;
     }
 
     @Override
-    public BotApiMethod<Message> handle(String messageFromUser, String chatId, State state)
+    public CommandResult handle(String messageFromUser, String chatId, State state) throws IOException
     {
+        String[] arguments = getSplitArguments(messageFromUser);
+        if (!checkArgumentsCount(2, arguments))
+        {
+            return new CommandResult(new SendMessage(chatId, ConstantManager.NO_FILE_NAME_FOUND), true);
+        }
+        if (!arguments[0].equals("/delete"))
+            return new CommandResult(new SendMessage(chatId, "Некорректная команда!"), false);
+        final String fileName = arguments[1];
         try
         {
-            String[] arguments = getSplitArguments(messageFromUser);
-
-            if (!checkArgumentsCount(2, arguments))
-            {
-                throw new IOException(ConstantManager.NO_FILE_NAME_FOUND);
-            }
-            final String fileName = arguments[1];
-            try
-            {
-                fileManager.deleteFile(fileName, chatId);
-                return new SendMessage(chatId, "Файл успешно удален.");
-            }
-            catch (IOException e)
-            {
-                throw new IOException("Не удалось удалить файл %s. ".formatted(fileName) + e.getMessage(), e);
-            }
+            fileManager.deleteFile(fileName, chatId);
+            return new CommandResult(new SendMessage(chatId, "Файл успешно удален."), true);
         }
-        catch (IOException exception)
+        catch (IOException e)
         {
-            return handleException(exception, chatId);
+            if (e.getMessage().equals(ConstantManager.NO_SUCH_FILE_EXISTS))
+                return new CommandResult(new SendMessage(chatId, "Не удалось удалить файл %s. ".formatted(fileName)
+                        + ConstantManager.NO_SUCH_FILE_EXISTS), true);
+            throw new IOException("Не удалось удалить файл %s. ".formatted(fileName) + e.getMessage(), e);
         }
     }
 }
